Fix installer
Ensure `exit 1` is only called in the curl and wget failure cases

Move `spaceship.zsh` into the oh-my-zsh folder, rather than using a
symbolic link. Since the /tmp folder is inherently temporary, when the
file is removed, the symbolic link will break.

Move the file extension for the `sed` command to directly after the `-i`
in order to allow Linux users to run this command as well. `sed` differs
slightly between macOS and Linux and Linux, which is unfortunate.