Add options to change secondary char and root char
This change allows users to decide if they want a different character for the
secondary prompt and for the root user.  This is achieved with the variables
`SPACESHIP_CHAR_SYMBOL_ROOT`, which is used in place of the default char
symbol if the current UID is 0, and `SPACESHIP_CHAR_SYMBOL_SECONDARY`, which
is used instead of the default char for the PS2 prompt.