KubeContext Section: Custom color based on context
Feature: add the ability to specify section color based on context names
matching a pattern.

Main use case is to mark contexts that should be handled with care, such
as "staging" or "production" environments.

An example of a use case might be to add the following to .zshrc:
```zsh
typeset -A SPACESHIP_KUBECONTEXT_COLORGROUPS=(
  ['dev-system']=green
  ['production']=red
  ['staging']=red
  ['^test-[0-9]+$']=yellow
  ['\.k8s\.local$']=red
)
```