Change kubecontext section to improve performance
Kubecontext will now display instantly for most cases.

Slower display is expected when changing kubectx or kubens.

Until now there were 2 calls to kubectl config, and that is not a cheap
call to make. Now it will only call kubectl once and determine the output based on
it's result.

The result is cached in a file so we can fetch it much faster than with
kubectl.

The assumption is that kubectl will always use a kubeconfig file to
determine it's active context, so we check the cached context against
the currently active $KUBECONFIG ( or the default $HOME/.kube/config)

Idea taken from jonmosco/kube-ps1.