Don't require aws-cli to exist when evaluating the AWS section
AWS_PROFILE is a cross-application standard and often used when
implementing software that uses AWS but does not rely on aws-cli,
eg. terraform, k8s or the AWS SDKs.
For projects that are built on top of those, it's very useful to
see the AWS_PROFILE section without having to install the aws-cli
first, especially since it serves no explicit purpose.