Respect ZDOTDIR env var when looking for .zshrc
The location of .zshrc is not always $HOME/.zshrc. The environment
variable ZDOTDIR defines in which directory .zshrc is in. Thus this
commit checks if ZDOTDIR is set and in that case use it to set the
ZSHRC environment variable. If ZDOTDIR is not set, $HOME is used as
fallback directory.
Additionally, it changes all hardcoded references related to this issue
to take ZDOTDIR into account.