feat(host): Add IPv4 address to the host section
The patch adds the IPv4 address to the host name if connected via SSH.

By default, the related SPACESHIP_HOST_SHOW_IP variable is `false` and
acts as before.

If the variable is set to `true`, it tries to figure out what IP to
show: it will query the ipv4 address of the first default gateway.

If the variable is set to any other string, it will be handled as the
interface to query and show the IPv4 address for.

resolve #1222